ci(semver-checks): avoid stale caches after rustc or tool upgrade
With an upgrade to the Rust toolchain, the version of the rustdoc JSON may change. We incorporate the rustc version into the cache key to automatically invalidate the cache in that case. Additionally, we also incorporate the version of the `cargo semver-checks` tool.

Related: https://github.com/obi1kenobi/cargo-semver-checks/issues/415.

Pull-Request: #3664.