ci(mergify): automatically approve PRs from maintainers
We have a limit of 1 approval for each PR in order to go into the merge queue.
To make life for us maintainers easier, we configure mergify to approve our and
only our PRs. This allows a single maintainer to land small PRs without having
to ping another maintainer.

Pull-Request: #3832.